BUG/MINOR: filters: Don't force the stream's wakeup when we wait in flt_end_analyze
In flt_end_analyze, we wait that the anlayze is finished for both the request
and the response. In this case, because of a task_wakeup, some streams can
consume too much CPU to do nothing. So now, this is the filter's responsibility
to know if this wakeup is needed.

This fix should be backported in 1.7.
(cherry picked from commit 2b553de5951b1a60e0c01775a2e7ef84b6472cb3)